Moved sidebar button test to Homepage tests + added Route order test in app
